Hotfix: fix 2 bugs (#205)
* hotfix(main): revert int -> float step()

* fix(ui-zoom): please for the love of god please fix this shit

* fix(ui-zoom): this should probably do

+++ src/main/java/cc/polyfrost/oneconfig/config/annotations/Number.java
@@ -44,7 +44,7 @@ public @interface Number {
 
     float max();
 
-    float step() default 1f;
+    int step() default 1;
 
     int size() default 1;
